Update pytest version (#51)
* Update pytest version to remove py dependency

Dependabot flagged 'py' as being vulnerable to a DoS. It likely doesn't affect us, but the recommendation is that 'pytest' removed dependency on 'py' in 7.2.0 so let's do that. 

This is discussed in a thread found here: https://github.com/pytest-dev/py/issues/287#issuecomment-1290407715

* Remove 'py'

py is a dependency of pytest so can be removed as of pytest version==7.2.0.